Move legacy client pool into client pool package
