+++ src/metron/clientpool/legacy/client_pool.go
@@ -1,4 +1,4 @@
-package legacyclientpool
+package legacy
 
 import (
 	"fmt"
