+++ src/metron/clientpool/legacy/combined_pool.go
@@ -1,4 +1,4 @@
-package legacyclientpool
+package legacy
 
 import (
 	"errors"
